Minor fixes to pick-species menu
Fixed an edge-case in the scrollbar code.

Fixed a mistake in the pick-species menu code that caused us to truncate the scroll position.

Removed debug messages from the Pokemon icon pool helper.

+++ src/lu/pick_species_menu.c
@@ -818,8 +818,8 @@ static void PaintListingContents(void) {
    
    SetLoadingSpinnerVisible(sMenuState->widgets.loading_spinner_sprite_id, FALSE);
    
-   u8 pos   = GetListingScrollPosition();
-   u8 count = sMenuState->listing.contents.count;
+   u16 pos   = GetListingScrollPosition();
+   u16 count = sMenuState->listing.contents.count;
    {
       struct LuScrollbar* scrollbar = &sMenuState->widgets.scrollbar;
       scrollbar->scroll_pos = pos;
